nut mua lai trang lsmuahang

+++ app/src/main/java/com/example/bookstore/Model/DonHangGui.kt
@@ -4,8 +4,8 @@ data class DonHangGui(
     val MaNguoiDung: Int,
     val MaKhuyenMai: Int?,
     val PhuongThucThanhToan: String,
-    val PhiVanChuyen: Int,
-    val TongTien: Int,
+    val PhiVanChuyen: Double,
+    val TongTien: Double,
     val DiaChiGiaoHang: String,
     val GhiChu: String?,
     val ChiTiet: List<ChiTietDonHang>
