+++ app/src/main/java/com/example/bookstore/Model/SachtrongGioHang.kt
@@ -2,10 +2,10 @@ package com.example.bookstore.Model
 
 data class SachtrongGioHang(
     val MaSach: Int,
-    val MaGioHang:Int,
+    val MaGioHang: Int,
     val TenSach: String,
     val TenTacGia: String,
-    val GiaBan: Int,
+    val GiaBan: Double,
     val AnhBia: String,
     val SoLuong: Int
 )
